Unpin `build-essential`
The [Haskell Dockerfile Linter][hadolint] (`hadolint`) lint recommends
([DL3008][DL3008]) that container dependencies should be pinned to a
specific version. However, these development containers are designed to
pull the latest version of each dependency.

Because these containers are designed not for development environments,
the builds do not have to produce consistent results. All that matters
is that once published, each containers provides an identical
environment for VS Code and GitHub Actions.

[DL3008]: https://github.com/hadolint/hadolint/wiki/DL3008
[hadolint]: https://github.com/hadolint/hadolint